add server-context.h
